test smallpt
temp

add smallpt test

+++ visualize/raytracer.cpp
@@ -22,7 +22,7 @@ void BVHTracer::trace()
 
     // Raytrace over every pixel
     #pragma omp parallel for
-    for (size_t i = 0; i< _width; ++i)
+    for (int i = 0; i< _width; ++i)
 	{
         for (size_t j = 0; j< _height; ++j)
 		{
